requirements.txt
asgiref==3.5.2
Django==4.0.8
django-filter==22.1
sqlparse==0.4.3
